Prepare Base Outline and Add Detail
This commit adds the base layout for the dashboard. It places everything in place, with dummy content for the cards in order to be worked on later. The layout was created using grid, and the items inside laid out using either grid or flexbox. Icons and fonts were added. The header and sidebar are close to completion, with remainder of the work to be continued in the main content. A search was added to the header, along with three buttons, a profile greeting, and icons. On the sidebar, a header was added, along with buttons plus their icons. Background colors were added as well.